requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0; platform_python_implementation != "PyPy"
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6; sys_platform == "win32" or platform_system == "Windows"
cryptography==42.0.8
distlib==0.3.8
dj-database-url==2.2.0
django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.15.1
flask==3.0.3
gunicorn==22.0.0
h11==0.14.0
identify==2.5.36
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markupsafe==2.1.5
netifaces==0.11.0
nodeenv==1.9.1
packaging==24.1
passage-identity==2.1.4
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycparser==2.22; platform_python_implementation != "PyPy"
pydotplus==2.0.2
pyjwt==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-magic==0.4.27; sys_platform == "linux"
python-magic-bin==0.4.14; sys_platform == "win32"
pyyaml==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1; sys_platform == "win32"
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
virtualenv==20.26.2
werkzeug==3.0.3
whitenoise==6.6.0
whitenoise[brotli]==6.6.0
